refactor: move state logic into Flyout component
